requirements.txt
cycler==0.10.0
grpcio==1.41.0
grpcio-tools==1.41.0
kiwisolver==1.3.2
matplotlib==3.4.3
mpl-interactions==0.19.0
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
protobuf==3.18.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
